action resolver, /start,/sign_up commands

+++ src/main/java/com/devdmitryn/userofthedaybot/repository/ChatRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatRepository extends CrudRepository<CustomChat,Integer> {
+public interface ChatRepository extends CrudRepository<CustomChat,Long> {
 }
