+++ src/main/java/com/devdmitryn/userofthedaybot/service/ChatService.java
@@ -5,6 +5,6 @@ import com.devdmitryn.userofthedaybot.entity.CustomChat;
 public interface ChatService {
 
     void saveOrUpdate(CustomChat chat);
-    CustomChat getById(int id);
-    void delete(int chatId);
+    CustomChat getById(long id);
+    void delete(long chatId);
 }
